Updates from 2022 previous for coding
Updated castvotes to save based on array the 1/2 step to get it to a relation database. This includes make the code proper not depending on a fixed array so we can drop a db solution if needed or move it to other solutions.
Added randomizer to it
Also Added checks and verification for save
Fixed spelling error around candidates.